fix: adds toJson method to dart mappable enums
Dart Mappable does not generate a toJson method for enums, instead generating a toValue method. There are no plans to change this. API client generators like Retrofit rely on detecting a toJson method to determine whether to use the enums name value or a serialized value.

This commit adds a toJson method to all enums.